+++ lib/i18n.ts
@@ -10,6 +10,7 @@ const STRINGS = {
     readArticle: "Lire l'article",
     chatTitle: "Arthur Mathon",
     closeChat: "Fermer la conversation",
+    newChat: "Démarrer une nouvelle conversation",
     letsConnect: "Restons en contact !",
     connectText:
       "N'hésitez pas à me contacter sur les réseaux ou à m'envoyer un e-mail.",
@@ -36,6 +37,7 @@ const STRINGS = {
     readArticle: "Read the article",
     chatTitle: "Arthur Mathon",
     closeChat: "Close the conversation",
+    newChat: "Start a new conversation",
     letsConnect: "Let's connect!",
     connectText:
       "Feel free to contact me on social media or send me an email.",